Build search object from archive. Sort search in vue.
