+++ src/main.js
@@ -129,6 +129,7 @@ Vue.mixin({
       }
     },
     $getResourceURL: Resources.getResourceURL,
+    $getChapter: Resources.getChapter,
     $filterURL(u) {return this.$getResourceURL(u)},
     $pushURL(to, key = this.$localData.tabData.activeTabKey){
       const url = this.$resolvePath(to)
