+++ src/resources.js
@@ -203,6 +203,100 @@ const UrlFilterMixin = {
   }
 }
 
+function getChapter(key) {
+  // Just putting this here because both processes need this logic.
+  let p = parseInt(key)
+  if (!p) {
+    switch (key) {
+      case 'jb2_000000': return 'Jailbreak'
+      case 'MC00001': return 'Blood Spade'
+      case 'pony': return 'Homestuck Act 3'
+      case 'darkcage': return 'Homestuck Act 6 Intermission 1'
+      case 'pony2': return 'Homestuck Act 2'
+      case 'darkcage2': return 'Homestuck Act 6 Act 3'
+    }
+  }
+  else if (p <= 135 && p >= 1) return 'Jailbreak'
+  else if (p <= 218 && p >= 136) return 'Bard Quest'
+  else if (p <= 1892 && p >= 219) {
+    let c = 'Problem Sleuth '
+    if (p >= 1841) c += "Epilogue"
+    else if (p >= 1708) c += "Chapter 22"
+    else if (p >= 1655) c += "Chapter 21"
+    else if (p >= 1589) c += "Chapter 20"
+    else if (p >= 1507) c += "Chapter 19"
+    else if (p >= 1466) c += "Chapter 18"
+    else if (p >= 1406) c += "Chapter 17"
+    else if (p >= 1299) c += "Chapter 16"
+    else if (p >= 1257) c += "Chapter 15"
+    else if (p >= 1149) c += "Chapter 14"
+    else if (p >= 1069) c += "Chapter 13"
+    else if (p >= 1030) c += "Chapter 12"
+    else if (p >= 953) c += "Chapter 11"
+    else if (p >= 873) c += "Chapter 10"
+    else if (p >= 816) c += "Chapter 9"
+    else if (p >= 742) c += "Chapter 8"
+    else if (p >= 666) c += "Chapter 7"
+    else if (p >= 604) c += "Chapter 6"
+    else if (p >= 546) c += "Chapter 5"
+    else if (p >= 448) c += "Chapter 4"
+    else if (p >= 402) c += "Chapter 3"
+    else if (p >= 302) c += "Chapter 2"
+    else if (p >= 219) c += "Chapter 1"
+    return c
+  }
+  else if (p <= 1900 && p >= 1893) return 'Homestuck Beta'
+  else if (p >= 1901) {
+    let c = 'Homestuck '
+    if (p >= 10027) c += "Act 7"
+    else if (p >= 9987) c += "Act 6 Act 6 Act 6"
+    else if (p >= 9349) c += "Act 6 Act 6 Intermission 5"
+    else if (p >= 9309) c += "Act 6 Act 6 Act 5"
+    else if (p >= 8844) c += "Act 6 Act 6 Intermission 4"
+    else if (p >= 8821) c += "Act 6 Act 6 Act 4"
+    else if (p >= 8801) c += "Act 6 Act 6 Intermission 3"
+    else if (p >= 8753) c += "Act 6 Act 6 Act 3"
+    else if (p >= 8431) c += "Act 6 Act 6 Intermission 2"
+    else if (p >= 8375) c += "Act 6 Act 6 Act 2"
+    else if (p >= 8178) c += "Act 6 Act 6 Intermission 1"
+    else if (p >= 8143) c += "Act 6 Act 6 Act 1"
+    else if (p >= 8092) c += "Act 6 Intermission 5 Intermission 6"
+    else if (p >= 8012) c += "Act 6 Intermission 5"
+    else if (p >= 8011) c += "Act 6 Intermission 5 Intermission 5"
+    else if (p >= 7965) c += "Act 6 Intermission 5 Intermission 4"
+    else if (p >= 7922) c += "Act 6 Intermission 5 Intermission 3"
+    else if (p >= 7882) c += "Act 6 Intermission 5"
+    else if (p >= 7881) c += "Act 6 Intermission 5 Interfishin"
+    else if (p >= 7866) c += "Act 6 Intermission 5"
+    else if (p >= 7847) c += "Act 6 Intermission 5"
+    else if (p >= 7846) c += "Act 6 Intermission 5 Intermission 2"
+    else if (p >= 7840) c += "Act 6 Intermission 5"
+    else if (p >= 7839) c += "Act 6 Intermission 5 Intermission 1"
+    else if (p >= 7823) c += "Act 6 Intermission 5"
+    else if (p >= 7688) c += "Act 6 Act 5 Act 1 x2 Combo"
+    else if (p >= 7678) c += "Act 6 Act 5 Act 1"
+    else if (p >= 7614) c += "Act 6 Act 5 Act 2"
+    else if (p >= 7412) c += "Act 6 Act 5 Act 1"
+    else if (p >= 7341) c += "Act 6 Intermission 4"
+    else if (p >= 7338) c += "Act 6 Act 4"
+    else if (p >= 7163) c += "Act 6 Intermission 3"
+    else if (p >= 6720) c += "Act 6 Act 3"
+    else if (p >= 6567) c += "Act 6 Intermission 2"
+    else if (p >= 6320) c += "Act 6 Act 2"
+    else if (p >= 6195) c += "Act 6 Intermission 1"
+    else if (p >= 6013) c += "Act 6 Act 1"
+    else if (p >= 6011) c += "Intermission 2"
+    else if (p >= 4526) c += "Act 5 Act 2"
+    else if (p >= 3889) c += "Act 5 Act 1"
+    else if (p >= 3258) c += "Act 4"
+    else if (p >= 3054) c += "Intermission"
+    else if (p >= 2660) c += "Act 3"
+    else if (p >= 2149) c += "Act 2"
+    else c += "Act 1"
+    return c
+  }
+}
+
 module.exports = {
   init(settings){
     assets_root = settings.assets_root || assets_root
@@ -215,5 +309,6 @@ module.exports = {
   resolveURL,
   resolvePath,
   getResourceURL,
+  getChapter,
   resolveAssetsProtocol
 }
